12 Oct 2004 Tweak cckd free space processing - Greg Smith 12 Oct 2004 Fix broken suspend/resume - Greg Smith
git-svn-id: file:///home/jj/hercules.svn/trunk@2450 956126f8-22a0-4046-8f4a-272fa8102e63

+++ sr.h
@@ -521,7 +521,7 @@ do { \
 #define SR_READ_HDR(_file, _key, _len) \
 do { \
   size_t _rc; \
-  size_t _buf[8]; \
+  BYTE _buf[8]; \
   _rc = SR_READ(_buf, 1, 8, (_file)); \
   if (_rc != 8) SR_READ_ERROR; \
   (_key) = fetch_fw(_buf); \
